requirements.txt
pandas>=1.5.0
boto3>=1.26.0
python-dateutil>=2.8.2
python-dotenv>=1.0.0
hubspot-api-client>=8.0.0 
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.10